refactor: API 호출 로직과 tanstack-query 훅 분리 및 type 분리
