+++ src/api/bundle/api.ts
@@ -1,10 +1,15 @@
 import { BUNDLE_COLORS } from "@/constants/constants";
-import { GiftBox, ReceiveBundle } from "@/types/bundle/types";
+import {
+  GiftBox,
+  PutCharacterPayload,
+  PutCharacterResponse,
+  ReceiveBundle,
+  ResultGiftBox,
+} from "@/types/bundle/types";
 import { getCookie } from "cookies-next";
 import { PICKTORY_API } from "../api-url";
 
 /** 보따리 생성 api */
-
 export const createBundle = async ({
   bundleName,
   selectedBagIndex,
@@ -44,7 +49,6 @@ export const createBundle = async ({
 };
 
 /** 보따리 업데이트 api */
-
 export const updateBundle = async (giftBoxes: GiftBox[]) => {
   const accessToken = getCookie("accessToken");
 
@@ -91,8 +95,122 @@ export const updateBundle = async (giftBoxes: GiftBox[]) => {
   return await response.json();
 };
 
-/** 보따리 풀어보기 api */
+/** 보따리 마저 채우기 api */
+export const fetchFillGift = async (bundleId: number) => {
+  const accessToken = getCookie("accessToken");
 
+  if (!bundleId) return;
+
+  const response = await fetch(PICKTORY_API.getDraftBundles(bundleId), {
+    method: "GET",
+    headers: {
+      Authorization: `Bearer ${accessToken}`,
+      Accept: "application/json",
+    },
+  });
+
+  if (!response.ok) {
+    throw new Error("데이터를 불러오는 데 실패했습니다.");
+  }
+
+  return response.json();
+};
+
+/** 배달부 설정 api */
+export const putDelivery = async ({
+  bundleId,
+  deliveryCharacterType,
+}: PutCharacterPayload): Promise<PutCharacterResponse> => {
+  const accessToken = getCookie("accessToken");
+
+  const response = await fetch(PICKTORY_API.putBundleDelivery(bundleId), {
+    method: "PUT",
+    headers: {
+      Authorization: `Bearer ${accessToken}`,
+      Accept: "application/json",
+      "Content-Type": "application/json",
+    },
+    body: JSON.stringify({ deliveryCharacterType: deliveryCharacterType }),
+  });
+
+  if (!response.ok) {
+    throw new Error("데이터를 불러오는 데 실패했습니다.");
+  }
+
+  return response.json();
+};
+
+/** 메인화면 보따리 조회 api */
+export const fetchBundlePreview = async () => {
+  const accessToken = getCookie("accessToken");
+
+  const response = await fetch(PICKTORY_API.getBundlesPreview, {
+    method: "GET",
+    headers: {
+      Authorization: `Bearer ${accessToken}`,
+      Accept: "application/json",
+    },
+  });
+
+  console.log(response);
+
+  if (!response.ok) {
+    throw new Error("데이터를 불러오는 데 실패했습니다.");
+  }
+  return response.json();
+};
+
+/** 보따리 목록 조회 api */
+export const fetchBundles = async () => {
+  const accessToken = getCookie("accessToken");
+
+  const response = await fetch(PICKTORY_API.getBundles, {
+    method: "GET",
+    headers: {
+      Authorization: `Bearer ${accessToken}`,
+      Accept: "application/json",
+    },
+  });
+
+  if (!response.ok) {
+    throw new Error("데이터를 불러오는 데 실패했습니다.");
+  }
+  return response.json();
+};
+
+/** 보따리 간이 조회 api */
+export const fetchMyBundleDetail = async (id: number) => {
+  const accessToken = getCookie("accessToken");
+
+  const response = await fetch(PICKTORY_API.getBundleDetail(id), {
+    method: "GET",
+    headers: {
+      Authorization: `Bearer ${accessToken}`,
+      Accept: "application/json",
+    },
+  });
+
+  return response.json();
+};
+
+/** 보따리 삭제 */
+export const deleteBundle = async (bundleId: number) => {
+  const accessToken = getCookie("accessToken");
+
+  const response = await fetch(PICKTORY_API.deleteBundle(bundleId), {
+    method: "DELETE",
+    headers: {
+      Authorization: `Bearer ${accessToken}`,
+      Accept: "application/json",
+    },
+  });
+
+  if (!response.ok) {
+    throw new Error("보따리 삭제에 실패했습니다.");
+  }
+};
+
+/** 보따리 풀어보기 api */
 export const fetchResponseBundle = async (link: string) => {
   try {
     const response = await fetch(PICKTORY_API.openBundle(link), {
@@ -120,20 +238,10 @@ export const fetchResponseBundle = async (link: string) => {
   }
 };
 
-/**보따리 결과 확인하기 api */
-
-export interface GiftData {
-  id: number;
-  name: string;
-  link: string;
-  thumbnail: string;
-  purchaseUrl: string;
-  responseTag: string;
-}
-
 /** 보따리 결과 조회 */
-
-export const fetchGiftResults = async (id: number): Promise<GiftData[]> => {
+export const fetchGiftResults = async (
+  id: number,
+): Promise<ResultGiftBox[]> => {
   const accessToken = getCookie("accessToken");
 
   const response = await fetch(PICKTORY_API.getBundleResult(id), {
@@ -149,5 +257,28 @@ export const fetchGiftResults = async (id: number): Promise<GiftData[]> => {
   }
 
   const jsonData = await response.json();
+
   return jsonData.result.gifts;
+};
+
+/** 보따리 개별 선물 조회 api */
+export const fetchGiftDetail = async (giftId: number, bundleId: number) => {
+  const accessToken = getCookie("accessToken");
+
+  const response = await fetch(
+    PICKTORY_API.getBundleResultDetail(bundleId, giftId),
+    {
+      method: "GET",
+      headers: {
+        Authorization: `Bearer ${accessToken}`,
+        Accept: "application/json",
+      },
+    },
+  );
+
+  if (!response.ok) {
+    throw new Error("데이터를 불러오는 데 실패했습니다.");
+  }
+
+  return response.json();
 };
