+++ src/queries/useDeleteMyBundle.ts
@@ -1,24 +1,5 @@
-import { PICKTORY_API } from "@/api/api-url";
 import { useMutation, useQueryClient } from "@tanstack/react-query";
-import { getCookie } from "cookies-next";
-
-/** 보따리 삭제 */
-
-const deleteBundle = async (bundleId: number) => {
-  const accessToken = getCookie("accessToken");
-
-  const response = await fetch(PICKTORY_API.deleteBundle(bundleId), {
-    method: "DELETE",
-    headers: {
-      Authorization: `Bearer ${accessToken}`,
-      Accept: "application/json",
-    },
-  });
-
-  if (!response.ok) {
-    throw new Error("보따리 삭제에 실패했습니다.");
-  }
-};
+import { deleteBundle } from "@/api/bundle/api";
 
 export const useDeleteBundle = () => {
   const queryClient = useQueryClient();
